Move nunjucks and markdown configuration into separate files
